move wallet utils to shared package
