+++ apps/web/src/components/CommunityHolderGrid/CommunityHolderGridItem.tsx
@@ -13,8 +13,8 @@ import TokenDetailView from '~/scenes/TokenDetailPage/TokenDetailView';
 import { useReportError } from '~/shared/contexts/ErrorReportingContext';
 import { CouldNotRenderNftError } from '~/shared/errors/CouldNotRenderNftError';
 import getVideoOrImageUrlForNftPreview from '~/shared/relay/getVideoOrImageUrlForNftPreview';
+import { graphqlTruncateUniversalUsername } from '~/shared/utils/wallet';
 import { getOpenseaExternalUrl } from '~/utils/getOpenseaExternalUrl';
-import { graphqlTruncateUniversalUsername } from '~/utils/wallet';
 
 import HoverCardOnUsername from '../HoverCard/HoverCardOnUsername';
