+++ apps/web/src/components/ManageWallets/ManageWalletsRow.tsx
@@ -7,8 +7,8 @@ import { BaseM, BODY_MONO_FONT_FAMILY } from '~/components/core/Text/Text';
 import { walletIconMap } from '~/components/WalletSelector/multichain/WalletButton';
 import useRemoveWallet from '~/components/WalletSelector/mutations/useRemoveWallet';
 import { ManageWalletsRow$key } from '~/generated/ManageWalletsRow.graphql';
+import { graphqlTruncateAddress } from '~/shared/utils/wallet';
 import { isWeb3Error } from '~/types/Error';
-import { graphqlTruncateAddress } from '~/utils/wallet';
 
 import breakpoints from '../core/breakpoints';
 import { HStack } from '../core/Spacer/Stack';
